[LSTM/GRU Cell opt] Separate ref and opt versions. Replace the sigma / tanh function call with a direct activation function call.
